requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
streamlit
pyyaml
imbalanced_learn
streamlit